python-flask/requirements.txt
Flask==0.10.1
python-dotenv==0.3.0
